WIP Working on Vector Layer commands and changes. When the analyst changes the style color from the accordian, it will change both the stroke and fill colors correctly.
When the analyst changes the fill color, it does not properly persist/restore that color change. It might have something to do with the differences between config['color'], config['fillColor'], and config['fill']['color'] values.

There are no changes yet to Feature Actions. Feature Edits and Feature Commands (as opposed to Vector Layer commands) should be done and working properly.